feat: Add HomeViewModel and IHomeViewModel interfaces
This commit adds the `HomeViewModel` class and the `IHomeViewModel` interface to the project. The `HomeViewModel` is a `ViewModel` subclass that will be used for the home screen of the app. The `IHomeViewModel` interface defines the required properties and functions for the `HomeViewModel` class.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or functionality being added.